Refactor: Content Dto 수정
