+++ src/main/java/com/java/everyboard/content/dto/ContentAllResponseDto.java
@@ -1,10 +1,11 @@
 package com.java.everyboard.content.dto;
 
-import com.java.everyboard.content.contentEnum.Category;
+import com.java.everyboard.constant.Category;
 import lombok.AllArgsConstructor;
 import lombok.Builder;
 import lombok.Getter;
 
+import javax.persistence.Column;
 import javax.validation.constraints.NotBlank;
 import javax.validation.constraints.NotNull;
 import java.time.LocalDateTime;
@@ -25,7 +26,13 @@ public class ContentAllResponseDto {
     @NotNull
     private Category category; // Enum으로 교체될 예정
     private String tag;
-    private LocalDateTime createdAt;
-    private LocalDateTime modifiedAt;
+    private LocalDateTime createdAt; // 컨텐츠 작성시간
+    private LocalDateTime modifiedAt; // 컨텐츠 수정 시간
 
+
+
+    // 게시글 작성 유저 정보 //
+    private String nickname;
+    private String profileUrl; // 프로필 사진
+//    private List<CommentResponseDto> comments;
 }
