+++ src/main/java/com/java/everyboard/content/dto/UserContentResponseDto.java
@@ -17,7 +17,12 @@ public class UserContentResponseDto {
     private Long userId;
     @NotBlank(message = "게시글 제목을 입력해야 합니다.")
     private String title;
+    private String content;
     private LocalDateTime createdAt;
     private LocalDateTime modifiedAt;
+
+    // 유저 정보 //
+    private String nickname;
+    private String profileUrl; // 프로필 사진
 
 }
